requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
sqlalchemy==2.0.36
aiosqlite==0.20.0
argon2-cffi==23.1.0
aiosmtplib==3.0.2
itsdangerous==2.2.0
slowapi==0.1.9
filelock==3.16.1
pyyaml==6.0.2
pydantic-settings==2.7.1
email-validator==2.2.0
